print_pager.c: make it build with Unsquashfs
Remove the Mksquashfs specific BAD_ERROR() call and use
the generic Mksquashfs/Unsquashfs ERROR().

There is nothing that is being done at this point in Mksquashfs
that needs the additional handling of BAD_ERROR().

Print_pager.c is now a dependency of Unsquashfs due to the
compressor help routines it doesn't use.  However, in the future
Unsquashfs will be using the routines in print_pager.c to print the
help text, and so this isn't going to be an issue.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>