xz_wrapper, xz_wrapper_extended: Add RISC-V BCJ filter support
Set the LZMA2 option lp=1 by default with the RISC-V filter for
2-byte-aligned instructions. Sometimes also pb=1 would help but
it varies so this keeps the default pb=2.

RISC-V instructions are 2-byte aligned when the C extension is used
(the 16-bit compressed instructions). The assumption is that usually
the C extension is used. For 4-byte-aligned instructions lp=2,lc=2
would be better.

The RISC-V filter requires liblzma >= 5.6.0.

Linux 6.12 will hopefully include the RISC-V filter.[1]

Link: https://lore.kernel.org/lkml/20240721133633.47721-13-lasse.collin@tukaani.org/ [1]
Signed-off-by: Lasse Collin <lasse.collin@tukaani.org>