xz_wrapper, xz_wrapper_extended: Add ARM64 BCJ filter support
Set LZMA2 options pb=2, lp=2, lc=2 as the default with the ARM64 filter
as those are good for ARM64 code (all instructions are 4 bytes and
4-byte aligned). pb=2 is a default from the preset. lc=2 is needed
because the sum lc+lp must not exceed 4 and the default is lc=3.

The ARM64 filter requires liblzma >= 5.4.0. The code will build against
older liblzma but then the filter won't be available.

Linux 6.12 will hopefully include the ARM64 filter.[1]

Link: https://lore.kernel.org/lkml/20240721133633.47721-12-lasse.collin@tukaani.org/ [1]
Signed-off-by: Lasse Collin <lasse.collin@tukaani.org>