Mksquashfs: Remove single destination file fd and mutex
Previously the main thread and the writer thread shared the
same single fd to the destination file.  Because it was shared
both threads had to obtain a mutex and perform a lseek() to
their read or write position.

The writer thread now opens its own fd to the destination
file, which removes the contention and the necessary lock.
It also doesn't issue an lseek() if the current position
is the correct position in the file.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>