Revert "Mksquashfs: signal writer thread to close its fd."
This reverts commit 8804c8a90b8650617f5d238224cf2b851e6b2512.

There is a pathological situation where this commit is broken.

If the last file is a duplicate which needed blocks spilled to disk
(because the write block cache wasn't large enough to hold all the
blocks), a NULL block pointer will be queued to the writer thread, to
synchronise the main thread with the writer thread.  The destination
file will then be truncated to remove the blocks spilled to disk.

If this is the last file then no additional write blocks are sent to
the writer, and the main process will proceed to send two NULLs to
the writer thread to close fd and exit.

In this case three NULLs are sent in succession.  The first two will
write the fd and terminate the writer thread, and the third will
cause the main thread to hang waiting for an response from the
dead writer process.