mksquashfs: abstract getting of start position in write_file_XXX()
Previously, the code got the current write position (which is the start
of any written blocks) on entry to the routine.

After all blocks have been written it then passes that to the appropriate
routines to create the inode and to keep track of file blocks for duplicate
checking.

However, as far as the write routines are concerned they only need to know
this start value at the end.  So abstact this away from the write
routines themselves.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>