mksquashfs: on error, only truncate if blocks written
When writing a file with multiple blocks, if an I/O error occurs, the
file is skipped, and any blocks already written to disk need to be
removed by truncating the file.

Previously, the code always truncated, even if no blocks had been written.

The pos handling code now tracks the write position, and handles putting
of blocks to the writer thread, and it knows if any blocks have been written.

So add a reset_pos() function that resets the write position if necessary
and returns TRUE.  But if no resetting is necessary, return FALSE.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>