pseudo: add new 'h' definition which follows symlinks
The 'l' pseudo definition previously didn't follow symlinks, which
meant if a symlink was given as the filepath, a reference will be
created to the symlink and not to the file to which it points.

This commit adds a new 'h' definition which follows symlinks,
restoring the original behaviour of the 'l' definition which
doesn't follow symlinks.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>