mksquashfs: recursively scan directory tree and create linked list
In the reader thread (responsible for reading files from disk), the
directory tree is recursively scanned, and files are read whilst
scanning the directory tree.

This commit changes this to create a linked list of files when the
directory tree is recursively scanned.  This linked list of files is
then walked and the files read at this point.

The reason for doing this is to enable in the future:

1. Parallel reading of files on multiple threads,
2. Additional read-ahead threads to eliminate pipe-line stalls.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>