reader: sequence buffers from reader thread using file_count & blocks
Rather than sequencing buffers from the reader thread using a single
sequence counter, switch to using a file_count and block count.

Currently, if a file needs to be re-statted and re-read (because
it has changed size), the file_count is incremented to ensure
the re-read blocks are sequenced differently.  When files
are read in parallel in the future this will need to change.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>