caches-queues-lists: implement queue_get_tid()
Previously, a thread can go to sleep if the queue is empty.

Now, a thread can also get to sleep if there are no available idle cores.

To implement this, two additional calls are added:

1. Wait_thread_idle() - the calling thread may go to sleep in this function
   waiting for an idle core.

2. Set_thread_idle() - this tells the system that this deflator thread is
   idle.  The core is now free for any threads waiting for a core.

If a thread sets its state to idle, because it is waiting for a buffer,
it obviously has to re-execute wait_thread_idle() to wait for an idle
core again.

Currently wait_thread_idle() and set_thread_idle() are stubs.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>