thread: implement wait_thread_idle() and set_thread_idle()
Wait_thread_idle() will cause the caller to go to sleep if
there's no idle cores.  Fragment deflators are highest priority
(because they're compressing the oldest data and Mksquashfs
has to wait for this to compress before writing a multi-block
file), and so they are always guaranteed cores (up to all of them).
Readahead data blocks are given any remaining idle cores.

Also fix get_thread_id() to use total * 2, as there are double the
number of deflators than cores.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>