reader: move linked list to (re-)alloced array
It was initially decided to store the reader 'work list' of files to be
read in a linked list.  Parallel reader threads would scan this linked
list either consuming files from the current head, or files within the
list depending on the scan criteria.  Each thread would take their
chosen item from the list, and adjust it accordingly, so that other
threads would not use it.

However, this list manipulation will have to be done with a mutex
held, and it has been decided that this may introduce lower performace
due to thread contention on this lock.  The more reader threads the more
likely this will happen.

This commit changes the code to use an array, with the intention that
parallel scanning and updating of the array can be without taking a
lock.  This should be more scalable.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>