reader: add version to sequencing
Buffers were sequenced by file count and buffer count.  But, if a file
changes size, and needs to be re-read, then this isn't sufficent,
because the re-read file buffers will have the same sequencing.

This commit fixes this by introducing a version that is bumped every
time the file needs to be re-read.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>