deflator: copy version from reader buffer to writer buffer
The writer buffer is passed to the main thread via the reader sequenced
queue, and it needs version to be set.

Also set version to 0 in tar.c.  Version isn't used in the tar code,
but, it needs to be set to 0 for the reader sequenced queue and the
write file code.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>