mksquashfs.c: add symbolic mode support for -force-file-mode
Symbolic modes are based on Chmod symbolic modes.

If Mode is Symbolic, permissions can be Set, Added or Removed.

The symbolic mode is of the format [ugoa]*[[+-=]PERMS]+
	PERMS = [rwxXst]+ or [ugo]
	and the above sequence can be repeated separated with commas.

A combination of the letters ugoa, specify which permission bits will
be affected, u means user, g means group, o means other, and a
means all or ugo.

The next letter is +, - or =.  The letter + means add to the existing
permission bits, - means remove the bits from the existing permission
bits, and = means set the permission bits.

The permission bits (PERMS) are a combination of [rwxXst] which
sets/adds/removes those bits for the specified ugoa combination. They
can alternatively be u, g or o, which takes the permission bits from the
user, group or other of the file respectively.

Examples:

1. -force-file-mode u+r

Adds the read permission to user.

2. -force-file-mode ug+rw

Adds the read and write permissions to both user and group.

3. -force-file-mode u=rw,go=r

Sets the permissions to rw-r--r--, which is eqivalent to

  -force-file-mode 0644

5. -force-file-mode ug=o

Sets the user and group permissions to the permissions for other.

The old parse_mode() function has been renamed to parse_mode_old() and
it will removed once the other command options have been updated to
use the new parse_mode()/mode_execute() functions.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>