mksquashfs: add symbolic modes to -all-root and -default-root
If Mode is Symbolic, permissions can be Set, Added or Removed.

The symbolic mode is of the format [ugoa]*[[+-=]PERMS]+
	PERMS = [rwxXst]+ or [ugo]
	and the above sequence can be repeated separated with commas.

A combination of the letters ugoa, specify which permission bits will
be affected, u means user, g means group, o means other, and a
means all or ugo.

The next letter is +, - or =.  The letter + means add to the existing
permission bits, - means remove the bits from the existing permission
bits, and = means set the permission bits.

The permission bits (PERMS) are a combination of [rwxXst] which
sets/adds/removes those bits for the specified ugoa combination. They
can alternatively be u, g or o, which takes the permission bits from the
user, group or other of the file respectively.

Examples:

1. -all-root u+r

Adds the read permission to user.

2. -all-root ug+rw

Adds the read and write permissions to both user and group.

3. -all-root u=rw,go=r

Sets the permissions to rw-r--r--, which is eqivalent to

4. -all-root 0644

5. -all-root ug=o

Sets the user and group permissions to the permissions for other.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>