reader: add explict file_count to read_array entries
The single read_array will be split into separate arrays, one for
fragments (less than block size), and one for blocks (block size or
larger).  This makes it easier for multiple reader threads to scan
the arrays locklessly.

Before that an explicit file count has to be added, because once
split into separate arrays, the array index can't be used for the
global file count.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>