reader: add synchronisation information to file buffers
Previously sequenced queues were sequenced via a single field
(sequence), which was incremented every time a file buffer
was sent from the reader thread.

Due to the simplicity of this, the reader_queue_get() code could easily
work out what the next buffer to be returned was (sequence + 1), meaning
this did not have to done in the calling code.

When the buffer sequencing was moved to a 3 value triple of file_count,
block and version, it was no longer possible for reader_queue_get()
to work out what the next buffer to be returned was.  So the logic was
moved to the main thread that receives the buffers in sequence order.
Specifically the write_file() routine then explicitly called
set_next_file() and set_next_version() to tell the queue what was
the next buffer to be returned.

This has worked well.  But, it relies on the following restrictions:

1. The code that receives the buffers (and calls set_next_file() and
   set_next_version()) has to be intelligent and track the buffers to
   know when to call set_next_file() and set_next_version().

2. The code that receives the buffers has to be single threaded,
   otherwise additional complexity will need to be added to synchronise
   the threads.

It so happens that not only do the buffers have to sequenced on delivery
to the main thread, but they also have to be sequenced on delivery to
the multiple block compressor threads and the fragment preprocessor
threads, _before_ they are then re-sequenced for delivery to the main
thread.

Neither the block compressor threads nor the fragment preprocessor
threads conform to the above restrictions, they lack intelligence and
are multi-threaded.

So the logic has to be moved back into fragment_queue_get().  But
how since it lacks the intelligence to do this?

The alternative (and better) solution in this commit is to add
a flag to each buffer, indicating what the next buffer in the
sequence will be, i.e. it is the next block in the same file, the
first block in the next file, or a new version (if the file failed to
be read and it is making another attempt).

This additional information means the logic in fragment_queue_get() can
successfully work out what the next block is going to be.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>