mksquashfs: add -force-dir-mode as analogue of -force-uid and -force-gid
Options -force-uid and -force-uid were added many years ago in
Mksquashfs 2.0 (2004), but mode setting options were never added.

One reason is blanket mode setting is made complex because of the
different handling of mode bits between directories and non-directories
and also between executables and non-executables.

This commit adds a -force-dir-mode option which takes a symbolic mode
or an octal mode which only affects directories.  This solves the
directory/non-directory issue.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>