print_pager: always use tab_out in autowrap_print()
Always tab out every line printed with the value of tab_out.  That is
tab out every line when an embedded newline (\n) is found or the
line wraps.

Previously it only tabbed out a line if it wrapped, and not when an
embedded newline was found.

This allows text which has been pre-formatted with an embedded
newline to not also have to embed the tab out value.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>