caches-queues-lists: rename reader_queue to main_queue
The reader queue sequences data blocks arriving to the main thread,
and so rename it to main queue.

This allows the queues from the reader thread to the deflate threads
and the process_fragment threads to be named reader_queues. Currently
these queues are not sequencd, but, that needs to change once we have
parallel reader threads.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>