caches-queues-lists: implement a new sequenced queue
The data blocks read by the parallel reader threads are queued to
the parallel deflate (compression) threads.  The data blocks
received from the queue by the deflate threads should be in
sequence order, because the filesystem is obviously compressed
and constructed in sequence order.  In other words, the ahead of
time reads by the parallel reader threads to increase I/O
bandwidth should not be compressed ahead of time or added to
the filesystem ahead of time.

The same applies to fragment blocks (files less than one block
in size).  These are queued to the parallel fragment_processing
threads, and they should arrive in sequence order.

The problem is reader buffers are sequenced by a triple of
file_count, buffer_count and version.  This ensures data blocks
and fragment blocks are correctly sequenced when they arrive
at the main thread to construct the filesystem.  In theory when
these blocks are queued to the deflate threads and the
fragment_processing threads, this triple should also ensure that
they are correctly sequenced there. Unfortunately that is not the
case because the triple sequences will not be consecutive and the
sequenced queue implementation relies on consecutive sequences.

For example files 1 2 and 3 will get file_count 1, 2 and 3.  Blocks
from these files will all arrive at the main thread, and so the
file_counts are consecutive.  But imagine if file 1 and 3 are larger
than a block size, and file 2 is smaller than a block size.  In this
case the deflate threads will be queued file 1 and 3, and the
fragment_processing threads will be queued file 2.  In both cases
for each queue the file_count is no longer consecutive.

This commit solves this problem by adding a new sequenced queue
which does not rely on the triple sequences being consecutive.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>