reader: move resource creation to separate function
Not all the requested threads may be created because there may
not be any work for them (i.e. no fragments for a fragment reader
thread).

So resource creation should be done only after fragment_count,
block_count and total_count are known.

Obviously this will be when the final necessary code
changes are done.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>