Makefile: fix default option value with CONFIG enabled
The default for USE_PREBUILT_MANPAGES in this case should be "n",
and not 1.

In practice it doesn't matter, as "y" is taken as yes, and anything
else as no.  But it is wrong ...

Also the non CONFIG definition of USE_PREBUILD_MANPAGES should have
spaces around the assignment operator to make it consistent
with all the other definitions.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>