reader: add dump of active threads to the "queues and caches" dump
Now that there is a thread abstraction in Mksquashfs that tracks
active and inactive threads, this can be used to add a dump of
the active threads to the queues and caches dump.

Also use processors variable to calculate the total number
of threads, rather than calling get_nprocessors() which doesn't
take into acount the -processors option.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>