sqfstar: check isextended on bit level
GNU tar and libarchive check the isextended byte on bit level, not as
real number. The code currently only works because isextended being
-1 is still considered true and 0 (false) occurs because the octal
parsing ends at '\0' and takes the already parsed value.

This means that 0x00 becomes 0, everything else -1 and thus true.

The compatibility ends if the ASCII character '0' is inserted, which
sqfstar interprets as 0 (false), while GNU tar and libarchive treat
it as != 0x00 and thus true.

Signed-off-by: Samanta Navarro <ferivoz@riseup.net>