mksquashfs: don't memmove() in write_dir() for every 8K block
Compressing a very large directory (12 million entries or about 240 Mbytes),
it was discovered Mksquashfs was hanging in write_dir().  This is because
the routine unnecessarily memmoves() the cached directory data for every 8K
compressed, and given that the cached data starts out at 240 Mbytes,
that is a lot of memmoving!

This is ancient code dating back to 2002, when directories
were small, and it should have been optimised years ago.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>