mksquashfs: add environment variable SQFS_CMDLINE
I often see reports where people are trying to find out
what command line was passed to Mksquashfs when it is run
from a script.  This is often for debugging purposes when
something is going wrong, and also to know how the Squashfs
filesystem was built.

Many scripts don't offer the ability to dump the command line,
requiring the script to be modified, which is often beyond the
ability of the user or impossible if the script is unmodifiable
due to being read-only.

This commit adds an environment variable SQFS_CMDLINE, which if
set causes Mksquashfs to dump the command line to the directory
specified by SQFS_CMDLINE, with the file being called sqfs_cmdline.

The name of the file is deliberately set to sqfs_cmdline to prevent
people using this as a backdoor to overwrite an arbitrary file, which
if Mksquashfs is run with elevated permissions, they might normally
not be able to.

Signed-off-by: Phillip Lougher <phillip@squashfs.ork.uk>