mksquashfs: don't follow symbolic links in check_sqfs_cmdline
Refuse to open any pre-existing sqfs_cmdline file if
it is a symbolic link.  This is to prevent people trying
to write to an arbitrary file.

Also change to appending to the file.  This is done to
allow the cmdline of multiple invocations of Mksquashfs
to be captured, rather than the last Mksquashfs run.

It also makes it impossible to use this facility to over-write
the contents of an existing file.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>