mksquashfs: refuse to append to a hard link in check_sqfs_cmdline
If the sqfs_cmdline file has an nlink count greater than 1,
then there are multiple directory entries pointing to the
same file, and this is mostly likely because this is a hard
link to a pre-existing file (rather than the other way around).

This is an obvious red flag that someone is trying to get
Mksquashfs to write to an arbitrary file, and so refuse to
do so.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>