caches-queues-lists: add new write cache
Add a new specialised cache for the block deflate threads and
writer thread.  This is because the multiple block reader threads
can have their blocks compressed by the block deflate threads
at the same time, but, they can't all share the same writer
cache block allocation.

If you don't know why, think of the case where ahead of time blocks
are compressed by the block deflate threads and they take all the
available blocks, meaning any earlier reader thread timewise
will deadlock waiting for a free block that will never become
available.

Also (and more likely to happen) when duplicate checking is done,
the routine expects all the writer cache to be available, and if
blocks are taken by ahead of time blocks, then the routine can
deadlock for the previous reasons too.

So far the new write cache is a copy of the previous implementation
with a new api.  Later commits will rewrite and add the necessary
additional functionality.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>