reader: switch from is_fragment() to is_large()
We want all files less than a block size to be handled
by fragment reader threads, and files which are a block
size or larger to be handled by block reader threads.

The code was using the is_fragment() helper function
to do that, but, that is actually the wrong function because
it also takes into account things we're not interested in,
such as the -no-fragments option, which means files less
than a block size are not assigned to the fragment reader threads.
We don't care whether they're packed or not packed into a fragment,
we're simply interested in the pure size of the file.

Also this function will return TRUE if the file is a
multi-block file with a tailend that will be packed into
a fragment.  Again, we don't want the is_fragment()
function to return TRUE here, because it will asign it
to a fragment reader thread.

In the past although the above would have assigned files to the
wrong reader thread, it hasn't actually mattered because each
thread could handle the wrongly assigned file (this is because
they must if a file dynamically changes size).

Switch from is_fragment() to is_large() which simply returns
TRUE is a file is a block size or larger.  This does the
correct assignment.

Increase the write cache slot size allocated to fragment reader
threads from 0 to 1.  This is because with the -no-fragments
option, they need a write buffer allocation to store
the block deflated file.

A later commit will properly compute that value when dealing
with files that dynamically change size.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>