mksquashfs: avoid deadlock in write_file_blocks_dup()
Commit 214cc7be65d8fed840756a302b4e3467c247aaf0 reintroduces a potential
deadlock by reversing the order of buffer acquisition in the deflator
threads.

This deadlock is caused because getting the reader buffer and then
getting the writer buffer is not atomic.  Deflator X could get block
150, and before it gets a writer buffer, deflator Y could read block
151 and then get a writer buffer.

When duplicating checking a file in write_files_blocks_dup(), if a
file has too many blocks to fit into the write cache (now the write
cache slot), it has to spill the blocks that won't fit to disk.  For
example if the write cache has 100 blocks, and the file has 150 blocks,
the first 50 blocks are spilled to disk, leaving the last 100 blocks
using all of the write cache buffers.

Now, if the above deflator thread race happens when reading the last
block in the file (block 150) it will want to take the last cache
buffer, but another deflator takes the buffer for a later file. This
means the deflator thread and write_file_blocks_dup() will deadlock.

Swapping the order of buffer acquisition to write buffer first, ensures
the deflator thread can always get a reader buffer and compress it,
avoiding the race condition and deadlock.

The alternative approach which this commit implements is to remember
that no more than processors - 1 deflators can race, and so the
write_file_block_dup() routine can rely on max buffers less processor
buffers to be available.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>