reader: add more advice in check_min_memory()
If the amount of memory is too low there are ways of reducing it
rather than increasing memory.  But the ways very much depend
on the options selected.

For example if the readers are already single threaded, there's
not much point in asking the user to reduce them, and the error
messages will just look silly.

Likewise, if there's already only one processor in use, there's
not much use in asking the user to reduce it.

Also fix the issue that processors was not correctly set on
entry to check_min_memory(), and if a less than 1M value was
given to -mem, it would result in integer underflow, and all
zeros passed to check_min_memory().  This integer underflow
was previously protected by the minimum 16M value check.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>