mksquashfs: remove minimum physical memory check
This fixed value was never a good check.  The actual lowest usable
physical memory was always dependant on the options and the number of
processors.

In any case, the last time I ran Debian 12 in a virtual machine it
wouldn't run in less than 800M.  So such memory checks seem a moot
point now anyway.

The check_min_memory() function is much better, and will
do the right thing.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>