Only one remaining use of SQUASHFS_LOWMEM
There's now only one remaining use of SQUASHFS_LOWMEM, which is
where the amount of memory cannot be determined.

Rename and increase to a more sensible value of 1G, and let
check_min_memory() tell the user how to set the options to fit into
that amount of memory if necessary.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>