reader: handle files that dynamically change size
Most of this code is already in place, and this commit adds
the last bit, which to give each fragment reader thread
a write cache slot allocation of processors + 1.

A fragment reader thread is intended to read files less than
a block in size.  But, to accommodate the fact that a file may
dynamically change size at any time, it also has to be able to
handle files with mutiple blocks, or otherwise swap into a role
as a block reader thread.

To do this it needs a write cache slot allocation of its own, and
this needs to be at least processors + 1 (see commit
cd504573f5dfdc13beae245e6b7614f6b9594a65 for an explanation).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>