caches-queues-lists: create a special combined queue and cache
Create a special combined queue and cache that merges the
functionality of the read_queue and the write_cache.

Switch the code that uses the read_queue and write_cache
implementations to using the combined queue and cache.

The functionality and behaviour should be the same, except a
single mutex is used across the combined queue and cache, and
the data is in a single structure.

The reason for this should become clear in the next commit.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>