caches-queues-lists: don't unhash if not hashed
Previously in the deflate threads if a reader buffer was
all zero no writer buffer was obtained for it.  Now
a writer buffer is always obtained.

This means when a reader buffer is all zero, the writer
buffer is discarded unhashed, and put on the freelist
unhashed.

This breaks the previous situation when a buffer was
always hashed before being discarded, and put on to
the freelist.

So this commit adds a flag, and doesn't unhash the buffer
if it hasn't been hashed.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>