reader: increase maximum amount of readers to 8192:8192
For now you should ensure that the maximum number of reader
threads is below your maximum open file limit (run
ulimit -n), which is usually 1024.  Or of course you can
set the limit higher with ulimit -n.

If you do specifiy more readers than your open file limit
currently you'll get the mysterious error message

libgcc_s.so.1 must be installed for pthread_exit to work

The reason for this is described in this bug on the Python
bug tracker, and it's rather silly.

https://bugs.python.org/issue44434

Basically glibc pthread_exit() loads an unwind function
from libgcc_s.so.1 using dlopen().  If the process is out of
file descriptors then this will fail, and pthread_exit()
will abort.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>