mksquashfs: fix rare 17 year old bug
Recent pre-release testing of the new parallel reader code showed up a
discrepancy in the number of reported duplicates when compressing a
large filesystem (about 160Gbytes, 4.2 million files and 1.5 million
duplicates).

Occasionally the number of reported duplicates would be less than
expected by between 1 and 5 files, that is at worst 1,656,263 duplicates
rather than the correct 1,625,268 files. The effect of this is a correct
filesystem, but where 1 - 5 files have not been seen as duplicates and
so stored twice, which means a loss of compression.

The problem turns out to a bug introduced in Mksquashfs back in 2008,
in commit 325010a328655f026199af64fffda560d8e2bf4f (when the project
was still in CVS source control).  That commit added the ability in
duplicate file checking to look up a buffer in the cache, rather
than reading it back from the filesystem.  This was much faster than
previous for two reasons: first the duplicate checking code didn't
have to stall waiting for the buffer to be written to disk (it could
get the buffer when it was still waiting to be written to disk), and
secondly even after it was written to disk, it could get the buffer
from the cache if it was still there on the freelist waiting to be
reused, which would save a slower read back from the filesystem.

But there was an over-sight in the code.  To be able to look up the
buffers in the cache they had to put into a hash table indexed on
the byte location that the block (was/will be) written to in the
filesystem.  So far so good.  The problem was the code pre-emptively
hashed the buffers when the file was processed, and before the file was
determined to be a duplicate or a non-duplicate.

Now, if the file is determined to be a non-duplicate, the buffers
will be (at some point) written to the filesystem, and before that
they can be found in the cache.  But, what happens if the file is
found to be a duplicate?  The buffers in the cache are discarded and
never written to the filesystem.  But they are still hashed to
the location they would have been written to in the filesystem if they
were found to be a non-duplicate.

Now later files reuse buffers from the freelist, and write their blocks
to the filesystem if they are non-duplicate and they are hashed to the
location the blocks are written to in the filesystem.

These blocks are valid blocks in the filesystem, and they will be
looked up in file duplicate checking.  If they are found in the cache
they will be reused otherwise they will be re-read from the filesystem.

Now, think about the situation that blocks that were discarded being
duplicate could have been stored in the filesystem in exactly the
same place that later blocks were stored in the filesystem.  If that
happens we could have two blocks hashed to the same location in the
cache, the discarded block and the valid block.  Now if the valid
block gets reused without the discarded block being reused, it means a
cache lookup will retrieve the discarded block as a match for the index,
rather than either the correct block or no block.

If that happens then the file being duplicate checked will not match a
previously stored file that is a duplicate, resulting in a discrepancy
in the number of reported duplicates.

This is highly unlikely to occur relying on a train of unlikely
circumstances.  In fact to my knowledge I have never seen this bug until
the introduction of parallel reader threads.  But if about 10 parallel
reader threads are active, and given a large filesystem which randomly
hits the prerequisites, due to the extra parallelism they can trigger
this bug.

This commit fixes the essential over-sight in the code, which is buffers
are now not hashed until they are known to be non-duplicate, and
duplicate buffers are returned to the cache unhashed which means they
can't be looked up.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>