mksquashfs: go back to using single fd and locking
The main thread and the writer thread in Mksquashfs always shared a
single fd.  The writer thread obviously uses this for writing file
data blocks, but, the main thread also uses this to write out the
filesystem metadata.  More importantly as far as races is concerned,
it also uses this to read back data from the filesystem when
doing duplicate checking.

A couple of months ago, to eliminate the lock that has to be taken
around the single shared fd, I gave the writer its own fd.

Unfortunately extensive pre-release testing has shown up what looks to
be a race condition between the writer thread writing data on its fd, and
the main thread afterwards reading the data back on its fd, causing the
read to fail (EOF presumably)..  This has happened once, and I have been
unable to reproduce it.

Testing has also shown the removal of the lock doesn't improve
performance anyway.  So it is safer to back to using a single fd
rather than risk Mksquashfs aborting due to a race condition.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>