Makefile: specify number of parallel reader threads at build time
Allow the default number of reader threads to be changed at
build time.

Some systems may not benefit from parallel reader threads and
so that should not be the default, for example systems which
are known to have old fashioned mechanical hard drives.

On the otherhand, it may be known that Mksquashfs can benefit
from a massively parallel amount of threads, far in excess
of the small number given by default in the Makefile.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>