mksquashfs: additional sanity checks for single reader thread default
If the Mksquashfs default is a single reader thread, then require both
-small-readers and -block-readers options to be specified to change to
multiple readers.

The reason for this is because the number of small reader threads and
block reader threads is undefined with a single reader default.  So we
should force the user to specify how many they want of each.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>