mksquashfs: handle too large amount of readers specially
Explicitly tell the user that the thread count is too large, and give
the largest possible value.

Also if the total possible thread count exceeds their open file limit,
tell the user about that lower threshold too.

This means we should avoid annoying the user by complaining twice, first
that the value is too large, and then once fixed that it is still above
their open file limit.

Obviously we could avoid that by deliberately limiting the maximum
thread count to be below the typical process open file limit of
1024 (less margin of 10).  But first there may be systems which can
benefit from a larger amount of readers than 1014, and second the
open file limit could be lower than 1024, which means we still have
to deal with it anyway.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>