progressbar: print error messages on new line
If two error messages are printed within one second then the
second error message is printed on the same line as the progress
bar.

This is because the first error message disables and then re-enables
the progress bar. Re-enabling the progress bar does an explicit
display of the progress bar, but it doesn't set the need_nl flag.

Now normally the progress bar will be re-displayed within one second
within the SIGALARM handler, which does set the need_nl flag.

But, if a second error message occurs before that happens, then
the need_nl flag isn't set.

So enable_progress_bar() should set the need_nl flag.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>